Day_38_41/requirements.txt
boto3==1.21.32
certifi==2022.12.7
mysql-connector-python==8.2.0
numpy==1.21.5
opendatasets==0.1.22
pandas==1.3.5
python-dotenv==1.0.0
Requests==2.31.0
SQLAlchemy==1.4.39
